refactor: rename useWorkStore to useEditorStore
